+++ Documentos/Semana 9 - Base de versionamento e programação em Python/CalculadoraPython/pytest.ini
@@ -4,5 +4,3 @@ markers =
     subtracao: marca teste de operações de subtração
     multiplicacao: marca teste de operações de multiplicação
     divisao: marca teste de operações de divisão
-    potencia: marca teste de operações de potenciação
-    fatorial: marca teste de operações de fatorial
