+++ Documentos/Semana 9 - Base de versionamento e programação em Python/CalculadoraPython/tests/test_fatorial.py
@@ -1,0 +1,14 @@
+import calculadora
+import pytest
+
+def test_fatorial():
+    calc = calculadora.CalculadoraPython()
+    assert calc.fatorial(1) == 1
+    assert calc.fatorial(2) == 2
+    assert calc.fatorial(3) == 6
+    assert calc.fatorial(4) == 24
+    assert calc.fatorial(5) == 120
+
+def test_fatorial_de_zero():
+    calc = calculadora.CalculadoraPython()
+    assert calc.fatorial(0) == 1
