+++ README.md
@@ -1,9 +1,8 @@
 # <h1 align="center">AWS & AI for Software Quality Engineering</h1>
 
-<i><h4 align="center">Repositório com atividades realizadas na Bolsa de Estudos Compass.UOL</h4></i>
+<i><h4 align="center">Repositório com atividades feitas na Bolsa de Estudo Compass.UOL</h4></i>
 
 
-=======
 ---
 
 ## 📌 Informações
@@ -13,7 +12,6 @@
 - **Turma:** Jun25  
 
 ---
-  
   
 ## 📂 Organização
 
@@ -38,3 +36,4 @@ O conteúdo está dividido em **Sprints (Branches):**
 - **Sprint 5**
   - Semana 9 – Base de versionamento e programação em Python  
   - Semana 10 – Testes Automatizados + APIs REST com Robot Framework  
+
